+++ CameraControl.Devices/Nikon/NikonL830.cs
@@ -17,7 +17,7 @@ namespace CameraControl.Devices.Nikon
             ExposureMeteringMode = new PropertyValue<int> {Available = false};
             FNumber = new PropertyValue<long> {Available = false};
             IsoNumber = new PropertyValue<long> {Available = false};
-            CompressionSetting = new PropertyValue<int> {Available = false};
+            CompressionSetting = new PropertyValue<long> {Available = false};
             Mode = new PropertyValue<long>() {Available = false};
             ShutterSpeed = new PropertyValue<long>() {Available = false};
             WhiteBalance = new PropertyValue<long>() {Available = false};
